msm: kgsl: Remove L2PC PM QoS feature
Original commit modified and adapted to 4.19 // @itsHanibee

KGSL already has PM QoS covering what matters. The L2PC PM QoS code is
not only unneeded, but also unused, so remove it. It's poorly designed
anyway since it uses a timeout with PM QoS, which is drastically bad for
power consumption.

Change-Id: I3aba9f5c0cf09d8c5e13e5c5e87e20456ca1c5f4
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: itsHanibee <itshanibee@gmail.com>